feature: switch amount to float

+++ internal/transport/http/handler/v1/recipe/dto/common_body/ingredient.go
@@ -1,10 +1,10 @@
 package common_body
 
 type IngredientItem struct {
-	Id       string  `json:"id"`
-	Text     *string `json:"text,omitempty"`
-	Type     string  `json:"type"`
-	Amount   *int32  `json:"amount,omitempty"`
-	Unit     *string `json:"unit,omitempty"`
-	RecipeId *string `json:"recipeId,omitempty"`
+	Id       string   `json:"id"`
+	Text     *string  `json:"text,omitempty"`
+	Type     string   `json:"type"`
+	Amount   *float32 `json:"amount,omitempty"`
+	Unit     *string  `json:"unit,omitempty"`
+	RecipeId *string  `json:"recipeId,omitempty"`
 }
